Multiselect Issue of hitting php max_input_vars fixed

+++ application/models/DbTable/Enrollments.php
@@ -4,8 +4,8 @@ class Application_Model_DbTable_Enrollments extends Zend_Db_Table_Abstract
 {
 
     protected $_name = 'enrollments';
-    protected $_primary = array('scheme_id','participant_id');
-    
+    protected $_primary = array('scheme_id', 'participant_id');
+
     public function getAllEnrollments($parameters)
     {
 
@@ -13,8 +13,8 @@ class Application_Model_DbTable_Enrollments extends Zend_Db_Table_Abstract
          * you want to insert a non-database field (for example a counter or static image)
          */
 
-        $aColumns = array('p.unique_identifier','p.first_name','iso_name','s.scheme_name', "DATE_FORMAT(e.enrolled_on,'%d-%b-%Y')");
-	
+        $aColumns = array('p.unique_identifier', 'p.first_name', 'iso_name', 's.scheme_name', "DATE_FORMAT(e.enrolled_on,'%d-%b-%Y')");
+
         /*
          * Paging
          */
@@ -86,19 +86,19 @@ class Application_Model_DbTable_Enrollments extends Zend_Db_Table_Abstract
          * SQL queries
          * Get data to display
          */
-	
-	$sQuery = $this->getAdapter()->select()->from(array('p' => 'participant'))
-	                             ->join(array('c'=>'countries'),'c.id=p.country')
-                                     ->joinLeft(array('e'=>'enrollments'),'p.participant_id = e.participant_id')
-                                     ->joinLeft(array('s'=>'scheme_list'),'e.scheme_id = s.scheme_id',array('scheme_name' => new Zend_Db_Expr("GROUP_CONCAT(DISTINCT s.scheme_name ORDER BY s.scheme_name SEPARATOR ', ')")))
-				     ->where("p.status='active'")
-				     ->group("p.participant_id");
+
+        $sQuery = $this->getAdapter()->select()->from(array('p' => 'participant'))
+            ->join(array('c' => 'countries'), 'c.id=p.country')
+            ->joinLeft(array('e' => 'enrollments'), 'p.participant_id = e.participant_id')
+            ->joinLeft(array('s' => 'scheme_list'), 'e.scheme_id = s.scheme_id', array('scheme_name' => new Zend_Db_Expr("GROUP_CONCAT(DISTINCT s.scheme_name ORDER BY s.scheme_name SEPARATOR ', ')")))
+            ->where("p.status='active'")
+            ->group("p.participant_id");
 
         if (isset($sWhere) && $sWhere != "") {
             $sQuery = $sQuery->where($sWhere);
         }
         if (isset($parameters['scheme']) && $parameters['scheme'] != "") {
-            $sQuery = $sQuery->where("s.scheme_id = ? ",$parameters['scheme']);
+            $sQuery = $sQuery->where("s.scheme_id = ? ", $parameters['scheme']);
         }
 
         if (isset($sOrder) && $sOrder != "") {
@@ -122,12 +122,12 @@ class Application_Model_DbTable_Enrollments extends Zend_Db_Table_Abstract
 
         /* Total data set length */
         $sQuery = $this->getAdapter()->select()->from(array('p' => 'participant'), new Zend_Db_Expr("COUNT('p.participant_id')"))
-	                                    ->join(array('c'=>'countries'),'c.id=p.country')
-                                            ->joinLeft(array('e'=>'enrollments'),'p.participant_id = e.participant_id',array())
-					    ->joinLeft(array('s'=>'scheme_list'),'e.scheme_id = s.scheme_id',array())
-					    ->where("p.status='active'")
-					    ->group("p.participant_id");
-	
+            ->join(array('c' => 'countries'), 'c.id=p.country')
+            ->joinLeft(array('e' => 'enrollments'), 'p.participant_id = e.participant_id', array())
+            ->joinLeft(array('s' => 'scheme_list'), 'e.scheme_id = s.scheme_id', array())
+            ->where("p.status='active'")
+            ->group("p.participant_id");
+
         $aResultTotal = $this->getAdapter()->fetchAll($sQuery);
         $iTotal = sizeof($aResultTotal);
 
@@ -141,47 +141,47 @@ class Application_Model_DbTable_Enrollments extends Zend_Db_Table_Abstract
             "aaData" => array()
         );
 
-        
+
         foreach ($rResult as $aRow) {
             $row = array();
             $row[] = $aRow['unique_identifier'];
-            $row[] = $aRow['first_name']. " " .$aRow['last_name'];
+            $row[] = $aRow['first_name'] . " " . $aRow['last_name'];
             $row[] = $aRow['iso_name'];
             $row[] = $aRow['scheme_name'];
             $row[] = Pt_Commons_General::humanDateFormat($aRow['enrolled_on']);
-	    if(trim($aRow['scheme_name'])!=""){
-		$row[] = '<a href="/admin/enrollments/view/pid/' . $aRow['participant_id'] . '/sid/' . strtolower($aRow['scheme_id']) . '" class="btn btn-info btn-xs" style="margin-right: 2px;"><i class="icon-eye-open"></i> Know More</a>';
-	    }else{
-		$row[]="--";
-	    }
+            if (trim($aRow['scheme_name']) != "") {
+                $row[] = '<a href="/admin/enrollments/view/pid/' . $aRow['participant_id'] . '/sid/' . strtolower($aRow['scheme_id']) . '" class="btn btn-info btn-xs" style="margin-right: 2px;"><i class="icon-eye-open"></i> Know More</a>';
+            } else {
+                $row[] = "--";
+            }
 
             $output['aaData'][] = $row;
         }
 
         echo json_encode($output);
     }
-    
-    public function enrollParticipants($params){
-		
-	    $this->delete("scheme_id='".$params['schemeId']."'");
-		
-        foreach($params['participants'] as $participant){
-            $data = array('participant_id'=>$participant,'scheme_id'=>$params['schemeId'],'status'=>'enrolled','enrolled_on'=>new Zend_Db_Expr('now()'));
-            $this->insert($data);
+
+    public function enrollParticipants($params)
+    {
+
+        if (!empty($params['schemeId'])) {
+            $this->delete("scheme_id='" . $params['schemeId'] . "'");
+            $params['selectedForEnrollment'] = json_decode($params['selectedForEnrollment'], true);
+            foreach ($params['selectedForEnrollment'] as $participant) {
+                $data = array('participant_id' => $participant, 'scheme_id' => $params['schemeId'], 'status' => 'enrolled', 'enrolled_on' => new Zend_Db_Expr('now()'));
+                $this->insert($data);
+            }
         }
-		
-    }
-    
-    public function enrollParticipantToSchemes($participantId,$schemes){
-		
-		$this->delete("participant_id=".$participantId);
-		
-        foreach($schemes as $scheme){
-            $data = array('participant_id'=>$participantId,'scheme_id'=>$scheme,'status'=>'enrolled','enrolled_on'=>new Zend_Db_Expr('now()'));
-            $this->insert($data);
-        }
-		
     }
 
+    public function enrollParticipantToSchemes($participantId, $schemes)
+    {
 
+        $this->delete("participant_id=" . $participantId);
+
+        foreach ($schemes as $scheme) {
+            $data = array('participant_id' => $participantId, 'scheme_id' => $scheme, 'status' => 'enrolled', 'enrolled_on' => new Zend_Db_Expr('now()'));
+            $this->insert($data);
+        }
+    }
 }
